+++ packages/templates/tsconfig.json
@@ -4,6 +4,7 @@
     "outDir": "./dist",
     "declaration": false,
     "declarationMap": false,
+    "resolveJsonModule": true,
   },
   "exclude": ["dist", "build", "node_modules"],
   "extends": "@lightdotso/tsconfig/react.json",
